[FIX] 닉네임 수정시 정규식 적용 (#166)
