+++ src/main/java/com/project/trysketch/controller/ImageController.java
@@ -1,20 +1,11 @@
 package com.project.trysketch.controller;
 
-import com.project.trysketch.dto.request.UserRequestDto;
-import com.project.trysketch.entity.ImageLike;
 import com.project.trysketch.global.dto.DataMsgResponseDto;
-import com.project.trysketch.global.dto.MsgResponseDto;
 import com.project.trysketch.service.ImageService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
-import org.springframework.data.web.PageableDefault;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
 import javax.servlet.http.HttpServletRequest;
-import java.util.List;
 
 
 // 1. 기능   : ImageLike 컨트롤러
@@ -52,25 +43,6 @@ public class ImageController {
     public ResponseEntity<DataMsgResponseDto> imageLike(@PathVariable Long imageId,
                                                         HttpServletRequest request) {
         return ResponseEntity.ok(imageService.likeImage(imageId, request));
-    }
-
-//======================================================================================================
-    // 마이페이지에서 좋아요 누른 사진 조회
-    @GetMapping("/mypage/image-like") // 수정 추가 김재영 01.29
-    public ResponseEntity<Page<ImageLike>> getImage(@PageableDefault(size = 5,sort = "createdAt",direction = Sort.Direction.DESC) Pageable pageable,
-                                                    HttpServletRequest request) {
-        return ResponseEntity.ok(imageService.getImage(request, pageable));
-    }
-    // 마이페이지 회원조회
-    @GetMapping("/mypage")
-    public ResponseEntity<DataMsgResponseDto> getMyPage(HttpServletRequest request) {
-        return ResponseEntity.ok(imageService.getMyPage(request));
-    }
-
-    // 마이페이지 정보변경
-    @PatchMapping("/mypage/profile")
-    public ResponseEntity<DataMsgResponseDto> patchMyPage(@RequestBody UserRequestDto userRequestDto, HttpServletRequest request) {
-        return ResponseEntity.ok(imageService.patchMyPage(userRequestDto, request));
     }
