+++ src/main/java/com/project/trysketch/service/ImageService.java
@@ -1,8 +1,6 @@
 package com.project.trysketch.service;
 
-import com.project.trysketch.dto.request.UserRequestDto;
 import com.project.trysketch.dto.response.ImageLikeResponseDto;
-import com.project.trysketch.dto.response.UserResponseDto;
 import com.project.trysketch.global.dto.DataMsgResponseDto;
 import com.project.trysketch.global.dto.MsgResponseDto;
 import com.project.trysketch.global.exception.CustomException;
@@ -21,7 +19,6 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
 import javax.servlet.http.HttpServletRequest;
 import java.util.*;
 
@@ -179,46 +176,5 @@ public class ImageService {
         return new MsgResponseDto(StatusMsgCode.DELETE_IMAGE);
     }
 
-    // 마이페이지 회원조회
-    public DataMsgResponseDto getMyPage(HttpServletRequest request) {
 
-        // 유저 정보 가져오기
-        Claims claims = jwtUtil.authorizeToken(request);
-        User user = userRepository.findByEmail(claims.get("email").toString()).orElseThrow(
-                () -> new CustomException(StatusMsgCode.USER_NOT_FOUND)
-        );
-
-        // 유저 정보에서 필요한 정보( id, email, nickname, ImgUrl ) 추출
-        UserResponseDto userResponseDto = UserResponseDto.builder()
-                .id(user.getId())
-                .email(user.getEmail())
-                .nickname(user.getNickname())
-                .imagePath(user.getImgUrl())
-                .build();
-
-        return new DataMsgResponseDto(StatusMsgCode.OK,userResponseDto);
-    }
-
-    // 마이페이지 회원 닉네임, 프로필사진 수정
-    @Transactional
-    public DataMsgResponseDto patchMyPage(UserRequestDto userRequestDto, HttpServletRequest request) {
-
-        // 유저 정보 가져오기
-        Claims claims = jwtUtil.authorizeToken(request);
-        User user = userRepository.findByEmail(claims.get("email").toString()).orElseThrow(
-                () -> new CustomException(StatusMsgCode.USER_NOT_FOUND)
-        );
-
-        // 유저 프로필, 닉네임 변경
-        user.update(userRequestDto.getNickname(), userRequestDto.getImgUrl());
-
-        // 유저 정보에서 필요한 정보( id, email, nickname, ImgUrl ) 추출
-        UserResponseDto userResponseDto = UserResponseDto.builder()
-                .id(user.getId())
-                .nickname(user.getNickname())
-                .imagePath(user.getImgUrl())
-                .build();
-
-        return new DataMsgResponseDto(StatusMsgCode.UPDATE_USER_PROFILE, userResponseDto);
-    }
 }
